requirements.txt
streamlit==1.32.0
google-generativeai==0.4.1
groq==0.4.2
python-dotenv==1.0.1
sqlalchemy==2.0.27
pandas==2.2.1
requests==2.31.0
pydantic==2.6.3
